Add: Color.swift 생성, Color 타입 외부 파일로 이동
